refactor(model) weather관련 type 폴더 호출
